pyproject.toml
[tool.poetry]
name = "ricebot"
version = "0.1.0"
description = ""
authors = ["degenRobot <86513395+degenRobot@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"
openai = "^1.43.0"
fastapi = {extras = ["standard"], version = "^0.112.2"}
chromadb = "^0.5.5"
together = "^1.2.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
